[GOOSE][APP] Extract score logic to its own object
To ease maintainability and extensibility, we move score logic
in a separate object from the main loop, and rely on callbacks
to display state changes.

In addition to the current score of the player, we maintain a
"pending" score, corresponding to the dice roll until we know
whether we accept or reject it.
While, for now, it is kept compatible wit the previous behavior,
the objective will be to allow events to accept/reject the roll,
so more interesting events could be created, such as "After a loss,
swap the after-roll position with the worst player. After a win,
swap the before-roll position with the best player."